Several corrections to the file
line 20: add the CIR image :kbd:`rautjarvi_aerial.tif` and containing the digitized lakes. --->  "and" should be "that is"

line 23: were created in 1994 with the situation almost 20 years ago.  --->  this should probably be not "ago" but " later. 
                                                                                                         Since the images are from 2012 and the shapefile is from 1994.
													 Comparing them would give you the situation almost 20 years later, not 
                                                                                                         that from 20 years ago !
																	
line 58: to identify forest stands a scale between 1:3 000  --->   "Sentence is far to long. Better to start a new line at 'A scale...

line 106: Open menu:`Vector -->                              --->  "menu:" should be :menuselection: for reasons of consistency

line 107: as the gui:`Target vector layer`.                 --->  "gui:" should be :guilabel: for reasons of consistency
 
line 108:  As gui:`Join vector layer` select                 --->  "gui:" should be :guilabel: for reasons of consistency
 
line 110:  In gui:`Output table` select gui:`Keep all records --->  "gui:" should be :guilabel: for reasons of consistency  (twice)
 
line 111: Click gui:`OK`.                                    --->  "gui:" should be :guilabel: for reasons of consistency
 
 Open the table of the new layer and order it 
 so that the forest stands with information for 
 the :guilabel:`Protection` attribute.              ---> add "are on top" to grammaticly finish the sentence correct