Update forest_maps.rst
row 193: you will do in the next ,  --->  "in the" is superfluous, don't need it. Space between "next" and ","

row 209: the whole forest area and you have already an they give you an idea of  ---> " you have already an" is not needed in this sentence
  
row 287: the corresponding feature from the :kbd:`atlas_coverage`. ----> add "layer" after "the" to be clearer