Remove SAGA algorithms help files
Around 30/230 algs are fully or partially described. Only listing GUI labels and parameters, which are already seen by the user in the dialog can not be consideredas helpful. Moreover, there are many SAGA algorithms that are missing in the list and the structure of the help files does not match SAGA's hence leading to 404 pages