phrasing
amended: "These five components of the QGIS interface are described in more detail in the following sections. Two more sections present keyboard shortcuts and context help." to "Scroll down for detailed explanations of these features. " i felt that the second sentence was confusing as it wasnt obvious below - i may go back and change this if found otherwise - happy to be challenged