List available units for measurement tool
Closes #1741

For the area measurement, should "map units" not properly be called "square map units"?

Also, I think it would make sense to explain what "map units" refers to. I thought it would be whatever linear units the project CRS uses, but when I tried this out using QGIS 3.2, the measurements were also done in meters with a geographic coordinate system (where I would have expected the linear measurement to be done in degrees). Is this the expected behaviour?