Reset of the initial dataset
It happens that the structure of osm layers output by QuickOSM has changed recently and for some layers, it is not compatible with shapefile format (eg long field name).
Some important features had also moved away from the dataset, hence the addtion of a new layer
Also, because we try to use gpkg, the whole dataset is provided in both formats without being used necessarily in the exercises. With these changes, there's no more duplicate data.

Last but not least, the screenshots are updated based on the aforementioned "rules" and using a Ubuntu environment instead of Windows.