Improve the discovery of the command line options (#3868)
by setting them as subsection titles instead of simple paragraph whose rendering is the same as their text description. Incidentally, this also allows to directly link the option when referring to it in a discussion.
Limit the TOC depth at 2 to avoid the tens of options shown in the list.