Updates for QGIS version 3.4 and above (#4037)
I have provided a number of changes to allow the examples provide to work with QGIS version 3.4.7 and python 3. The use of str(quri.encodedUri(), "utf-8") appears to make line 284 functional but I remain unclear on which option is being recommended? The sole purpose of this documentation appears to be to show how to use the authentication manager but the comment below the example seems to imply uri(False) should be used instead. Trouble is that I get a "Download of capabilities failed: Protocol "" is unknown" if I use that version.