Avoid spreading CRS handling explanation in different places
Instead of explaining logic of crs in different places (crs tabs in
both Options and Project properties dialog, and the dedicated Projection chapter),
let's move all the explanation and logic to the dedicated chapter.
Then use the CRS tabs sections to only describe the GUI, not how options may interact.