Remove the 'Last updated' sentence in the left column
It seemed that after setting the html_last_updated_fmt param in
conf.py automagically there is a 'last updated' string in the footer
of every page.

Though my aim was also to minimize the number of pages changed
(by only putting it on the frontpage index), I'm fine with this.