Removed provider_options transform_context
The provider_options transform_context is passed as the last parameter to uri.setDataSource(), I have checked both in the python console and in the source code and what I find is that the last parameter is for the primary_key of the database table. I only find one overload of this method, so unless there is another on the way, I think this is a mistake in the documentation.