reformatted, reduced parameter list, rewrote start
The changes are in response to the comments by @DelazJ. The code block was reformatted to  be more attractive, to remove references to version numbers, and to reduce the number of parameters listed. The introductory text before the code block was rewritten to better explain why it is helpful to have so many parameters.