Load PostGIS raster using manual URI string. (#5385)
* Load PostGIS raster using manual URI string.

Manual URI has to be used b/c set methods are written for geometry and not for raster. Therefore a description of the parameters is given.

* load URI from a dictionary per @elpaso

I have made these edits to load URI connection parameters for a PostgreSQL raster from a dictionary. 
Thank you for your comments @elpaso. Your code returns an empty string on my system (QGIS version: 3.12.2-București, QGIS code revision: 8a1fb33634). The docs for QgsProviderRegistry.encodeUri has a note stating "this function may not be supported by all providers, an empty string will be returned in such case".

* corrected formatting problem with dictionary

* Update docs/pyqgis_developer_cookbook/loadlayer.rst

Co-authored-by: Håvard Tveite <havard.tveite@nmbu.no>

* Combined syntax for 3.12 and 3.13+

* GDAL 'mode', designates union or tiled raster

* add PostGIS raster 'postgresraster' per @DelazJ

Uses the new 'postgresraster' metadata provider. Edits per suggestions by @DelazJ.

* reformatted, reduced parameter list, rewrote start

The changes are in response to the comments by @DelazJ. The code block was reformatted to  be more attractive, to remove references to version numbers, and to reduce the number of parameters listed. The introductory text before the code block was rewritten to better explain why it is helpful to have so many parameters.

* remove NULL parameters

* sql not where as parameter

* indent of comment

* Change NULL to None per Travis

Co-authored-by: Håvard Tveite <havard.tveite@nmbu.no>

* Changed NULL to None per TRAVIS

* Update docs/pyqgis_developer_cookbook/loadlayer.rst

* added import for QgsProviderRegistry

QgsDataSourceUri is already in the .. testsetup:: loadlayer

Co-authored-by: Håvard Tveite <havard.tveite@nmbu.no>