Update text and images in Lesson 4.2 "Labels" (#5463)
* Add .vscode to .gitignore

* Add preposition for better grammar: "Hovering over the layer"

* Fix answers: "Invert selection" not "Help" and "Fields" not "Source Fields"

Direct link to file in master branch that shows 3. |invertSelection|
in raw mode:
https://github.com/qgis/QGIS-Documentation/blob/master/docs/training_manual/introduction/overview.rst#-try-yourself-2

* Update text and images in Lesson 4.2 "Labels" (#4634)

Lesson 4.2 in the Training Manual was called "The Label Tool",
which appears to refer to an older tool that's no longer part of
the QGIS GUI. Now, Labels are controlled through tabs in the
Layer Styling panel or the Layer Properties dialog menu, just
like Symbology.

Run through the exercises and update all of the text, images,
and the answer key to represent the new QGIS GUI as of the
QGIS master branch (v3.13+)

Related issues: #4634

* Add newline at end of file

* Update one screenshot I missed in 4.2.6

* Switch from bulleted lists (*) to numbered lists (#.)

This makes the steps more obvious and easier to reference in
issue reports and discussion. I used Lesson 3.2 as my guideline
for how images and text should be nested inside the numbered
list, and what should still be bulleted (the 4.2.4 Try Yourself
section with "Check your results" left-justified rather than
nested, just like the 3.2.4 Try Yourself section).

I considered removing periods at the end of list items to match
the Writing Guidelines...

    Do not use punctuation after list items that only contain a single simple sentence

...but noticed that most of the other lessons (3.1, 3.2, etc.)
break this rule and use a period at the end of all list items,
so I left them in.

* Do not use periods after list items that only contain a single simple sentence

* Reorder sentence that includes button image

* Switch from :kbd:`word` to ``word`` to match writing guidelines

* Remote end of single simple sentence punctuation from an answer

* Use some :guilabel: tags to make Text, Buffer, and Color pop

* Update try yourself instructions for brevity

* Switch sequential answer to use numbered list instead of bullets

* Apply suggestions from code review

(mostly ``places`` instead of :guilabel:`places`)

Co-authored-by: Håvard Tveite <havard.tveite@nmbu.no>

* Change `Ok` to `OK` for consistency

* Some more ``places`` instead of :guilabel:``places``

* Even more ``places`` instead of :guilabel:`places`

* Update docs/training_manual/vector_classification/label_tool.rst

Co-authored-by: Håvard Tveite <havard.tveite@nmbu.no>