Update composer_map.rst
Added detail to the CRS setting for map grids, explaining that it is automatically linked to the map CRS (not the project CRS as previously suggested).  Also explained the method to set back to following the map CRS as this is not obvious.  You need to set it to blank again, but clicking the button opens a CRS dialogue, fortunately selecting a heading rather than a CRS provides the blank output required.