Update section on running qgis_mapserv.fcgi for the first time
The existing documentation is confusing about what to expect when running `/usr/lib/cgi-bin/qgis_mapserv.fcgi` for the first time.  Since QGIS 3.16.x, the server will produce a 400 response that suggests it is not working, when in fact it is.  See my answer here: https://gis.stackexchange.com/a/394171/181982