Updated CorePlugin Section
fixes #6775
Added GRASS GIS, SAGA and OTB to the CorePlugin Section and removed the
GPS Tools.
Also added an information since they have to be configurated to be
accessible. (What was in my eyes a little bit illogical with a core
plugin, nevertheless i think its a good idea to split them from the
processing plugin)