ArcGIS Rest Server
Started to work on #5290 and realized that there is no real documentaion
for the ArcGIS REST Servers. So i first added the Table in
https://docs.qgis.org/testing/en/docs/user_manual/introduction/browser.html?highlight=arcgis%20rest%20servers#tiles-and-web-services
and updated an image.

My idea is to document it in the https://docs.qgis.org/testing/en/docs/user_manual/introduction/browser.html?highlight=arcgis%20rest%20servers#tiles-and-web-services
section, in the subsection "Connectiong to web services" as a new item.

Thoughts?