Save QGIS project in PostgreSQL
Added a Tip to the PostGIS layers section.
It is documented in the "Working with Project Files" section (
https://docs.qgis.org/testing/en/docs/user_manual/introduction/project_files.html
) that one can save projects to a postgresql database, but i felt like
if the user is looking for postgis data information, its good to know
that you can also save project files there and that it is unlikely that
one will also check the "working with Project Files"-section in that
process.
Thoughts?