Missing dot and colon
Line 1313 : "at the same place" --> misses a dot at the end of the sentence. Added the dot, so it reads: "at the same place."
Line 1634 : "The guilabel:`Customization` dialog" missed a colon before "guilabel:" so it broke formatting. Added the missing colon, so all should be fine now with "The :guilabel:`Customization` dialog"