Edit in Place order and multi type removal
I made the suggested changes of removing the multi from the features description.
I also moved the In place line above any notes and Default Menu (this is where it was in exiting documentation). Also added "without M enabled" to "fix geometries" because the fix geometries drops M values so edit in place will not work on layers that have M enabled.