All of the remaining Edit in place edits additions and feature
All of the remaining Edit in place additions and feature descriptions.
I think that is all of them although the new Roundness algo has edit in place functionality but it hasn't been added to the documentation yet.
I am getting a error in the vectortable.rst saying that it doesn't have a checkbox. I see that there is a checkbox in the Substitutions definitions at the bottom of the other files but this also says don't add things here so I'm leaving that unresolved and hoping it will resolve itself.