New doc for point clouds
fixes also #6319

@DelazJ

Hi Harrisou,

here finally is the documentation for the point clouds. Martin Dobias and Philip have read, corrected and completed it so far and now we all think it's ok. There are only questions about parts that can be left out and there is only a reference/link instead:

- "Information Properties" and "Source Properties" (should be number 19.2.1 and 19.2.2) are similar to other layers;
- the "Customize Color Ramp Legend" section in the "Attributes by Ramp Renderer" chapter (19.2.3.1) is identical to "Customize Color Ramp Legend" in the "16.1. Raster Properties Dialog chapter" (I just copied the text).

Personally, I prefer the chapters to be included here as well, because something can change in the sections that are referenced and then the content is no longer appropriate. But you are probably the best person to assess what makes the most sense.

Greetings,
Jörn