align in-place docu with implementation
To be available for in-place editing, processing algorithms implement
the `bool supportInPlaceEdit(…)`-method.
As this is not dependent on the algorithms only working on feature
source level or not modifying fields, but stands on its own, it might
make it easier for users if we communicate what these algorithms do, not
why or why not they are available.