Gentle intro: Use table directives for captions
This commit uses reST `table` directives to "attach" captions to tables
in the "Gentle Introduction to GIS." The original captions followed the
tables as separate paragraphs. This update moves those captions into the
tables.

In places, the caption is shortened, and text from the original
caption is retained as a separate paragraph.

Cross references to tables are also updated to use Sphinx `:numref:`
roles.